fix: handling custom exp engine for bamboo
