+++ packages/spacecat-shared-rum-api-client/src/functions/traffic-acquisition.js
@@ -11,7 +11,6 @@
  */
 
 import { classifyTrafficSource } from '../common/traffic.js';
-import { fetch } from '../utils.js';
 
 function extractHints(bundle) {
   const findEvent = (checkpoint, source = '') => bundle.events.find((e) => e.checkpoint === checkpoint && (!source || e.source === source)) || {};
@@ -48,81 +47,8 @@ function transformFormat(trafficSources) {
   }));
 }
 
-/* c8 ignore start */
-/*
- * throw-away code for a single customer who customized the experimentation engine
- * this code will be removed once they start using the default exp engine
- *
- * this function fetches experiment manifests, then merges variants data into controls data
- *
- * ie:
- *
- * if the customer runs for an experiment where variants are as following:
- *   control: /
- *   challenger-1: /a1/
- *   challenger-2: /a2/
- *
- * then data for the `/a1/` and `/a2` are counted towards `/`'s data
- */
-async function mergeBundlesWithSameId(bundles) {
-  if (!bundles[0]?.url.includes('bamboohr.com')) return bundles;
-  const manifestUrls = [
-    ...new Set(bundles.flatMap((bundle) => bundle.events
-      .filter((e) => e.checkpoint === 'experiment')
-      .map((e) => e.source))),
-  ].map((experiment) => fetch(`https://www.bamboohr.com/experiments/${experiment}/manifest.json`));
-
-  const experiments = await Promise.all(manifestUrls);
-  const variants = (await Promise.all(experiments.map((e) => e.json().catch(() => {}))))
-    .filter((json) => json && Object.keys(json).length > 0)
-    .flatMap((json) => json.experiences?.data ?? [])
-    .filter((data) => data.Name === 'Pages');
-
-  const mapping = variants.reduce((acc, cur) => {
-    Object.entries(cur)
-      .filter(([k]) => !['Name', 'Control'].includes(k))
-      .forEach(([, v]) => {
-        acc[new URL(v).pathname] = new URL(cur.Control).pathname;
-      });
-    return acc;
-  }, {});
-
-  const variantPaths = Object.keys(mapping);
-
-  const getControlPath = (url) => {
-    const path = new URL(url).pathname;
-    if (variantPaths.includes(path)) return mapping[path];
-    return path;
-  };
-
-  const byIdAndPath = bundles.reduce((acc, cur) => {
-    const controlPath = getControlPath(cur.url);
-    const key = `${cur.id}-${controlPath}`;
-    if (!acc[key]) acc[key] = [];
-    if (variantPaths.includes(new URL(cur.url).pathname)) {
-      // eslint-disable-next-line no-param-reassign
-      cur.url = new URL(controlPath, cur.url).href;
-    }
-    acc[key].push(cur);
-    return acc;
-  }, {});
-
-  const merged = Object.entries(byIdAndPath).flatMap(([, v]) => {
-    let value = v;
-    if (v.length > 1) {
-      v[0].events.push(...v.slice(1).flatMap((bundle) => bundle.events));
-      value = [v[0]];
-    }
-    return value;
-  });
-
-  return Object.values(merged);
-}
-/* c8 ignore end */
-
 async function handler(bundles) {
-  const merged = await mergeBundlesWithSameId(bundles);
-  const trafficSources = merged
+  const trafficSources = bundles
     .map(extractHints)
     .map((row) => {
       const {
